dashboard and donations table use shared compoenent
